[refactor] Moved `isValidIPAddress` to its own TS file
