+++ src/NetscriptEvaluator.js
@@ -10,7 +10,7 @@ import {Script, findRunningScript,
 
 import {parse, Node}                        from "../utils/acorn";
 import {printArray}                         from "../utils/HelperFunctions";
-import {isValidIPAddress}                   from "../utils/IPAddress";
+import {isValidIPAddress}                   from "../utils/helpers/isValidIPAddress";
 import {isString}                           from "../utils/helpers/isString";
 
 var Promise = require("bluebird");
